test_graphene/requirements.txt
aiohttp==3.7.3
aniso8601==7.0.0
async-timeout==3.0.1
attrs==20.3.0
chardet==3.0.4
click==7.1.2
fastapi==0.63.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
idna==3.1
multidict==5.1.0
promise==2.3
pydantic==1.7.3
Rx==1.6.1
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
ujson==4.0.2
uvicorn==0.13.3
uvloop==0.15.1
yarl==1.6.3
